WIP: Rewrites using several Livewire components
