+++ src/Http/Livewire/ModuleDescription.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Uccello\ModuleDesigner\Http\Livewire;
+
+use Livewire\Component;
+use Uccello\ModuleDesigner\Support\Traits\StepManager;
+use Uccello\ModuleDesigner\Support\Traits\StructureManager;
+
+class ModuleDescription extends Component
+{
+    use StepManager, StructureManager;
+
+    protected $listeners = ['stepChanged', 'structureChanged', 'iconSelected'];
+
+    public function render()
+    {
+        return view('module-designer::livewire.module-description');
+    }
+
+    public function iconSelected($icon, $target)
+    {
+        $this->structure['icon'] = $icon;
+    }
+}
