+++ src/Support/TraitsOld/HasIcon.php
