+++ src/Support/TraitsOld/StructureAdapter.php
